Add files via upload
data_get.R is the r script to get the data from the json. dataset.R is the r script for cleaning the data. The CSV file called dataset.csv files is a huge dataset so we need to make a subset dataframe and sometimes we need to remove NA columns if possible.